+++ src/main/java/net/carcdr/ycrdt/YXmlElement.java
@@ -266,6 +266,47 @@ public class YXmlElement implements Closeable {
     }
 
     /**
+     * Gets the parent of this XML element.
+     * The parent can be either a YXmlElement or YXmlFragment.
+     *
+     * @return The parent node (YXmlElement or YXmlFragment), or null if this element has no parent
+     * @throws IllegalStateException if the XML element has been closed
+     */
+    public Object getParent() {
+        checkClosed();
+        Object result = nativeGetParent(doc.getNativePtr(), nativePtr);
+        if (result == null) {
+            return null;
+        }
+
+        // Result is Object[2] where [0] = Integer type, [1] = Long pointer
+        Object[] array = (Object[]) result;
+        int type = ((Integer) array[0]).intValue();
+        long pointer = ((Long) array[1]).longValue();
+
+        if (type == 0) {
+            // Element
+            return new YXmlElement(doc, pointer);
+        } else if (type == 1) {
+            // Fragment
+            return new YXmlFragment(doc, pointer);
+        } else {
+            throw new RuntimeException("Unknown parent type: " + type);
+        }
+    }
+
+    /**
+     * Gets the index of this element within its parent's children.
+     *
+     * @return The 0-based index within parent, or -1 if this element has no parent
+     * @throws IllegalStateException if the XML element has been closed
+     */
+    public int getIndexInParent() {
+        checkClosed();
+        return nativeGetIndexInParent(doc.getNativePtr(), nativePtr);
+    }
+
+    /**
      * Checks if this YXmlElement has been closed.
      *
      * @return true if this YXmlElement has been closed, false otherwise
@@ -334,4 +375,6 @@ public class YXmlElement implements Closeable {
     private static native long nativeInsertText(long docPtr, long xmlElementPtr, int index);
     private static native Object nativeGetChild(long docPtr, long xmlElementPtr, int index);
     private static native void nativeRemoveChild(long docPtr, long xmlElementPtr, int index);
+    private static native Object nativeGetParent(long docPtr, long xmlElementPtr);
+    private static native int nativeGetIndexInParent(long docPtr, long xmlElementPtr);
 }